Removed old bits of code
