+++ src/common/components/file-drop-area/file-drop-area.test.tsx
@@ -1,14 +1,14 @@
 import { themeProvider } from '@app/tests/test-utils';
 import { render, screen } from '@testing-library/react';
-import InfoFile from './info-file';
+import FileDropArea from './';
 
-describe('info-file', () => {
+describe('file-drop-area', () => {
   it('should render component', () => {
     const mockSetFileData = jest.fn();
     const mockSetIsValid = jest.fn();
 
     render(
-      <InfoFile
+      <FileDropArea
         setFileData={mockSetFileData}
         setIsValid={mockSetIsValid}
         validFileTypes={['xlsx']}
